try 8-bit

+++ ex2.h
@@ -1,18 +1,18 @@
 #include <stdint.h>
 
 #define SOUND_DECLARE(name) \
-extern uint16_t _binary_##name##_raw_start[]; \
-extern uint16_t _binary_##name##_raw_end[];
+extern uint8_t _binary_##name##_raw_start[]; \
+extern uint8_t _binary_##name##_raw_end[];
 #define SOUND(name) {_binary_##name##_raw_start, _binary_##name##_raw_end}
 typedef struct {
-	const uint16_t *const start;
-	const uint16_t *const end;
-	const volatile uint16_t * pos;
+	const uint8_t *const start;
+	const uint8_t *const end;
+	const volatile uint8_t * pos;
 } Sound;
 extern volatile Sound sounds[8];
 
 /*
- * Declaration of peripheral setup functions 
+ * Declaration of peripheral setup functions
  */
 void setup_timer(uint16_t period);
 void setup_dac(void);
